Pagination Redesign
Pagination has now been moved to ext.menus
Customizable buttons
Support for string pages
Returns the sent message (Doesn't return interaction messages yet, library doesnt support it yet.
