+++ discord/ext/commands/__init__.py
@@ -17,4 +17,3 @@ from .converter import *
 from .cooldowns import *
 from .cog import *
 from .flags import *
-from .pagination import *
