+++ discord/ext/menus/__init__.py
@@ -1,0 +1,9 @@
+"""
+discord.ext.menus
+~~~~~~~~~~~~~~~~~~~~~
+
+An extension module to provide useful menu options.
+
+"""
+
+from .pagination import *
